fix if file not exist
git-svn-id: http://svn.ives.fr/svn-libs-dev/asterisk/fontventa/trunk@1647 4a32156a-8e32-406b-b38d-595c535062e7

+++ tools/updateWS.ksh
@@ -133,15 +133,18 @@ ConvertFile()
 
 SaveCurrentWaitScreen()
 {
-    printLine "Save current wait screen ... "
-    cp -f $MediaRep$MediaName $MediaRep$SaveMediaName
-    ret=$?
-    if [ $ret -ne 0 ] 
-        then 
-        PrintFailed
-    else 
-        PrintOK
-    fi
+   if [ -f $MediaRep$MediaName ] 
+     then 
+       printLine "Save current wait screen ... "
+       cp -f $MediaRep$MediaName $MediaRep$SaveMediaName
+       ret=$?
+       if [ $ret -ne 0 ] 
+           then 
+           PrintFailed
+       else 
+           PrintOK
+       fi
+   fi
 }
 
 RestorCurrentWaitScreen()
